% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecciones.ar.2019.R
\docType{data}
\name{medios_sim_leg_nac}
\alias{medios_sim_leg_nac}
\title{Bancas por agrupación}
\format{Un data frame con 3,665,201 filas y 7 variables:
\describe{
  \item{CODIGO_PROVINCIA}{Código de provincia}
  \item{PORCENTAJE_CARGA}{porcentaje de carga}
  \item{CODIGO_CATEGORIA}{Código de categoría de la elección}
  \item{CODIGO_AGRUPACION}{Código de agrupación}
  \item{BANCAS_AGRUPACION}{Bancas obtenidas}
  \item{OBSERVACIONES}{Observaciones}
}}
\source{
\url{https://resultados2019.gob.ar/resultados_detalle.zip}
}
\usage{
medios_sim_leg_nac
}
\description{
Un dataset conteniendo el detalle de las bancas obtenidas por agrupación,
y provincia derivado del archivo original medios_sim_leg_nac.dsv
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
