% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecciones.ar.2019.R
\docType{data}
\name{mesas_agrp_politicas}
\alias{mesas_agrp_politicas}
\title{Votos por agrupación}
\format{
Un data frame con 3,665,201 filas y 7 variables:
\describe{
  \item{CODIGO_DISTRITO}{Código de distrito}
  \item{CODIGO_SECCION}{Código de sección electoral}
  \item{CODIGO_CIRCUITO}{Código de circuito}
  \item{CODIGO_MESA}{Código de mesa}
  \item{CODIGO_CATEGORIA}{Código de categoría de la elección}
  \item{CODIGO_AGRUPACION}{Código de agrupación}
  \item{CODIGO_LISTA}{Código de lista}
  \item{VOTOS_AGRUPACION}{Cantidad de votos por agrupación}
}
}
\source{
\url{https://resultados2019.gob.ar/resultados_detalle.zip}
}
\usage{
mesas_agrp_politicas
}
\description{
Un dataset conteniendo el detalle de los totales de votos por mesa, abierto hasta la agrupación,
derivado del archivo original mesas_totales_agrp_politica.dsv
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
