% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecciones.ar.2019.R
\docType{data}
\name{mesas_totales}
\alias{mesas_totales}
\title{Mesas}
\format{Un data frame con 2,532,824 filas y 7 variables:
\describe{
  \item{CODIGO_DISTRITO}{Código de distrito}
  \item{CODIGO_SECCION}{Código de sección electoral}
  \item{CODIGO_CIRCUITO}{Código de circuito}
  \item{CODIGO_MESA}{Código de mesa}
  \item{CODIGO_CATEGORIA}{Código de categoría de la elección}
  \item{CONTADOR}{Desconocido}
  \item{VALOR}{Desconocido}
}}
\source{
\url{https://resultados2019.gob.ar/resultados_detalle.zip}
}
\usage{
mesas_totales
}
\description{
Un dataset conteniendo el detalle de mesas derivado del archivo
original mesas_totales.dsv
}
\details{
(ESTRUCTURA ORIGINAL)
}
\keyword{datasets}
